+++ src/screens/CreateTransactionStack/TransactionConfirmationScreen.tsx
@@ -15,7 +15,7 @@ import DebitCreditOverview from 'components/screens/CreateTransactionStack/Debit
 import UpdateButton from 'components/screens/CreateTransactionStack/UpdateButton';
 import Container from 'components/screens/CreateTransactionStack/Container';
 import {isEmpty} from 'lodash';
-import useMutateCreateTransaction from 'hooks/mutations/TransactionConfirmation/useMutateCreateTransaction';
+import useCreateTransaction from 'hooks/mutations/TransactionConfirmation/useCreateTransaction';
 
 type TransactionConfirmationScreenProps = {
   navigation: RootStackNavigationProp;
@@ -28,8 +28,7 @@ export default function TransactionConfirmationScreen({
   const {transaction, balance, accounts, deleteAccount, clearTransaction} =
     useTransaction();
 
-  const {createTransaction, isPending, isSuccess} =
-    useMutateCreateTransaction();
+  const {createTransaction, isPending, isSuccess} = useCreateTransaction();
 
   useEffect(() => {
     if (isSuccess) {
