+++ src/types/dtos/category.ts
@@ -1,0 +1,9 @@
+export type CategoryListDTO = {
+  primaryCategoryId: number;
+  primaryCategoryContent: string;
+  secondaryCategoryId: number;
+  secondaryCategoryContent: string;
+  tertiaryCategoryId: number;
+  tertiaryCategoryContent: string;
+  isDefault: boolean;
+}[];
